# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_project-platypus-perry_base_internal_model.User:
    properties:
      DateOfBirth:
        type: string
      Email:
        type: string
      ID:
        type: string
      Name:
        maxLength: 50
        minLength: 2
        type: string
      Password:
        minLength: 8
        type: string
      Phone:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: "#/definitions/gorm.DeletedAt"
      id:
        type: integer
      updatedAt:
        type: string
    required:
      - Email
      - Name
      - Password
      - Phone
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A RESTful API service providing user management and health check endpoints
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Base API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
        - application/json
      description: Check if the API service is up and running
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
        - health
  /users:
    post:
      consumes:
        - application/json
      description: Creates a new user with the provided information
      parameters:
        - description: User object that needs to be created
          in: body
          name: user
          required: true
          schema:
            $ref: "#/definitions/github_com_project-platypus-perry_base_internal_model.User"
      produces:
        - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: "#/definitions/github_com_project-platypus-perry_base_internal_model.User"
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
        - users
  /users/{id}:
    delete:
      consumes:
        - application/json
      description: Deletes a user by their UUID
      parameters:
        - description: User UUID
          format: uuid
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "400":
          description: Invalid UUID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
        - users
    get:
      consumes:
        - application/json
      description: Retrieves a user by their UUID
      parameters:
        - description: User UUID
          format: uuid
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: User found successfully
          schema:
            $ref: "#/definitions/github_com_project-platypus-perry_base_internal_model.User"
        "400":
          description: Invalid UUID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
        - users
    patch:
      consumes:
        - application/json
      description: Updates an existing user's information
      parameters:
        - description: User UUID
          format: uuid
          in: path
          name: id
          required: true
          type: string
        - description: Updated user object
          in: body
          name: user
          required: true
          schema:
            $ref: "#/definitions/github_com_project-platypus-perry_base_internal_model.User"
      produces:
        - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: "#/definitions/github_com_project-platypus-perry_base_internal_model.User"
        "400":
          description: Invalid UUID format or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
        - users
schemes:
  - http
  - https
swagger: "2.0"